move physics & entropy to contexts dir
