+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export { EntropicController } from "./entropic-controller"
-export { Entropy, useEntropyContext } from "./entropy"
+export { Entropy, useEntropyContext } from "../contexts/entropy"
 export { Field } from "./field"
 export { Matter } from "./matter"
-export { Physics, usePhysics } from "./physics"
+export { Physics, usePhysics } from "../contexts/physics"
 export { Particle } from "./particle"
 export { Title } from "./title"
